chore: extract package loading to BuildUtils
